% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasInAcetaminophenStudies.R
\docType{package}
\name{BiasInAcetaminophenStudies-package}
\alias{BiasInAcetaminophenStudies}
\alias{BiasInAcetaminophenStudies-package}
\title{BiasInAcetaminophenStudies: Quantifying Bias in Epidemiological Studies on the Association Between Acetaminophen and Cancer}
\description{
Quantifying Bias in Epidemiological Studies on the Association Between Acetaminophen and Cancer.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
